Moved classes needed for token verifcation to sercuirty package
