+++ pnnl.goss.core/core-api.bnd
@@ -1,4 +1,4 @@
 Export-Package: \
 	com.northconcepts.exception,\
 	pnnl.goss.core
-Bundle-Version: 10.0.4.${tstamp}
+Bundle-Version: 10.0.5.${tstamp}
