+++ pnnl.goss.core/security-ldap.bnd
@@ -1,3 +1,3 @@
 Private-Package:  \
 	pnnl.goss.core.security.ldap
-Bundle-Version: 1.0.129.${tstamp}
+Bundle-Version: 1.0.130.${tstamp}
