+++ pnnl.goss.core/src/pnnl/goss/core/security/SecurityConfig.java
@@ -2,7 +2,7 @@ package pnnl.goss.core.security;
 
 import java.util.Set;
 
-import pnnl.goss.core.security.jwt.JWTAuthenticationToken;
+import pnnl.goss.core.security.impl.JWTAuthenticationToken;
 
 public interface SecurityConfig {
 	public String getManagerUser();
