+++ pnnl.goss.core/src/pnnl/goss/core/security/packageinfo
@@ -1,1 +1,1 @@
-version 6.0.0
+version 5.0.0
